Version 0.2: Github gem.
Added a gemspec, the way github wants it. This should make the gem installable from github.
The gemspec can be re-generated using 'rake gem:gemspec'. This uses the template as its source.
The readme was updated to show how to use github to install the gem.
The todo and changelog were removed. They were not used.
Dependencies on other gems were removed from the code.